draft_augur_tree make boottree grab conditional

+++ tasks/tasks_nextstrain.wdl
@@ -1237,9 +1237,11 @@ task draft_augur_tree {
         cat /proc/loadavg > CPU_LOAD
         { cat /sys/fs/cgroup/memory/memory.max_usage_in_bytes || echo 0; } > MEM_BYTES
 
+        find -name "*.fasta.boottrees" > NAME
+        if [ -s NAME ] ; then
         find -name "*.fasta.boottrees" | xargs mv -t .
-        ls
         mv *.fasta.boottrees ~{out_basename}_~{method}.fasta.boottrees
+        fi
 
     >>>
     runtime {
